+++ src/LDL/Validators/Config/RegexValidatorConfig.php
@@ -16,13 +16,19 @@ class RegexValidatorConfig implements ValidatorConfigInterface
      */
     private $regex;
 
-    public function __construct(string $regex, bool $negated=false, bool $dumpable=true)
+    public function __construct(
+        string $regex,
+        bool $negated=false,
+        bool $dumpable=true,
+        string $description=null
+    )
     {
         RegexHelper::validate($regex);
 
         $this->regex = $regex;
         $this->_tNegated = $negated;
         $this->_tDumpable = $dumpable;
+        $this->_tDescription = $description;
     }
 
     /**