+++ src/LDL/Validators/Config/StrictClassComplianceItemValidatorConfig.php
@@ -15,7 +15,12 @@ class StrictClassComplianceItemValidatorConfig implements ValidatorConfigInterfa
      */
     private $class;
 
-    public function __construct(string $class, bool $negated=false, bool $dumpable=true)
+    public function __construct(
+        string $class,
+        bool $negated=false,
+        bool $dumpable=true,
+        string $description=null
+    )
     {
         if(!class_exists($class)){
             throw new \LogicException("Class \"$class\" does not exists");
@@ -24,6 +29,7 @@ class StrictClassComplianceItemValidatorConfig implements ValidatorConfigInterfa
         $this->class = $class;
         $this->_tNegated = $negated;
         $this->_tDumpable = $dumpable;
+        $this->_tDescription = $description;
     }
 
     /**