+++ src/LDL/Validators/StringValidator.php
@@ -12,9 +12,9 @@ class StringValidator implements ValidatorInterface
      */
     private $config;
 
-    public function __construct(bool $negated=false, bool $dumpable=true)
+    public function __construct(bool $negated=false, bool $dumpable=true, string $description=null)
     {
-        $this->config = new Config\BasicValidatorConfig($negated, $dumpable);
+        $this->config = new Config\BasicValidatorConfig($negated, $dumpable, $description);
     }
 
     public function validate($value): void