+++ CHANGELOG.md
@@ -8,6 +8,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ### Added
 
+- feature/1200373038347529 - Add description to validator config / Add human validator dumper
 - feature/1200240911349077 - Create OR & AND Validator chains, add "negate" capability in validator config, remove old generic ValidatorChain
 - feature/1200200918895360 - Get succeeded, error and lastExecuted validators in ValidatorChain
 - feature/1200117701811974 - Add more validators